requirements.txt
streamlit
tensorflow
pillow
numpy
huggingface-hub
opencv-python-headless
matplotlib
python-docx
fastapi
uvicorn
python-multipart
jinja2
